// web/src/lib/enums.ts
export enum HttpMethod {
  GET = "GET",
  POST = "POST",
  DELETE = "DELETE"
}

export enum HttpStatus {
  OK = 200,
  NO_CONTENT = 204,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  BAD_GATEWAY = 502
}

export enum CacheMode {
  NO_STORE = "no-store"
}

export enum ChatRole {
  USER = "user",
  ASSISTANT = "assistant"
}